+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Don't allow line breaks in URLs ([#1488](https://github.com/latex-lsp/texlab/issues/1488))
+
 ## [5.24.0] - 2025-11-03
 
 ### Added