+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.25.1] - 2025-12-30
+
+### Fixed
+
+- Fix detecting dependencies in projects with `latexmkrc` file ([#1520](https://github.com/latex-lsp/texlab/issues/1520))
+
 ## [5.25.0] - 2025-12-28
 
 ### Added